backend/requirements.txt
fastapi
uvicorn
python-multipart


langchain
langchainhub
langchain-community

google-ai-generativelanguage
google-api-python-client

pdfplumber
python-docx
sentence-transformers